[BUGFIX] Apply server timezone to datetime properties in record API
DateTime objects initialized with @ will be forced to use UTC,
but want all dates (however they are stored, as localtime native
field or UTC timestamp field) to be initialized as localtime.

Apply the same algorithm as already used by FormEngine to resolve
database values to `\DateTime` instances. Extract the logic into a
general purpose `DateTimeFactory` and use this factory in
DatabaseRowDateTimeFields. The factory is intended to be rolled out
in all places where database values are converted into `\DateTime`
representation.

Releases: main, 13.4
Resolves: #106453
Related: #98045
Change-Id: I10a68595e3596b4d79b4cccf4ae490c885fc971d