[BUGFIX] Use DateTimeFactory for database datetime value interpretation
Fix display of native datetime fields in TCA record titles when
timezone is not UTC and apply `DateTimeFactory::createFromTimestamp()`
where the `(new DateTimeImmutable())->setTimestamp()` approach has been
used to workaround the PHP issue that direct `DateTime` creation from
timestamps does not respect current PHP timezone. The usage of the
`DateTimeFactory::createFromTimestamp()` wrapper hopefully helps to make
the intent of this chain clear and to prevent recreating the same
issue in new code as in #106453.

Extbase DateTime resolution will be address with a separate patch
and is marked as a TODO for the time being.

Releases: main, 13.4
Resolves: #96890
Related: #98045
Related: #103693
Related: #106453
Change-Id: I92c3fb86b1267821336988756f8bba966568af82