[BUGFIX] Load JavaScript importmaps of custom settings types
Ensure that the JavaScript importmap configuration of custom settings
type providers is loaded inside the settings editor by preloading
all settings types, that will be rendered inside the editor.
This also improves performance of the initial rendering and allows
to drop the inline spinner, which is actually just visible for
a few milliseconds and therefore actualy more distracting than helpful.

Resolves: #107038
Releases: main, 13.4
Change-Id: I5701cf15a724ddf189425324bc36ae0287e9014b