[BUGFIX] Re-encode URLs that get passed to FormEngine from LinkBrowser
This is a workaround because InputTransformer.toSearchParams for some
reason decodes search parameters after assembling them. This causes
properly encoded URLs to be changed, which violates RFC 1738 section
2.2 "All unsafe characters must always be encoded within a URL." [1]

While we are now double-encoding due to the unnecessary decode, this
workaround is scoped to the reported issue. A followup patch will
address the decoding inside `AjaxRequest(...).withQueryArguments`.

[1]: https://www.rfc-editor.org/rfc/rfc1738#section-2.2

Resolves: #107475
Releases: main, 13.4
Change-Id: I046eda901a8fbd64d227a67d67ca60e13ef05828