[WIP][DEMO][FEATURE] Add MCP Server demo based on Actions API
Implement a simple MCP server demo to showcase that the Actions
API AI-tools adapter can be used to drive an MCP Server.

The MCP server exposes actions from the Actions API as MCP tools
and is implemented via Actions API itself.
There's a lot missing, like a filtering for which actions
shall be exposed in which context. This change is here to prepare
the Actions API design to be compatible with different downstream
needs (REST, MCP, OpenAI Tools, and so on…), without requiring
every functionality to be wrapped for REST and MCP manually.

Releases: main
Resolves: #
Change-Id: I2f2ff276cb533422b6aa6393a16d1ea7c9bf55ef